refactor: decompose campaign execution
